SelectionFunctions.Examples

Description

Existential quantifier for Cantor space was defined with help from this blog post by Martin Escardo http:math.andrej.com20070928seemingly-impossible-functional-programs/

Examples for Nat

fiN :: K Nat

Existential quantifier for set {2,4,6,8,10}.

p11 :: Integral a => a -> Bool

Some predicates on Nat.

p12 :: t -> Bool

p13 :: t -> Bool

p14 :: (Num a, Ord a) => a -> Bool

Examples for Baire

fiNN :: K Baire

Existential quantifier for the set of three functions in Baire.

fiCantor :: K Baire

Existential quantifier for Cantor space in Baire.

p21 :: (Eq a, Num a1, Num a) => (a1 -> a) -> Bool

Some predicates on Baire.

p22 :: (Num a1, Num a, Ord a) => (a1 -> a) -> Bool

p23 :: t -> Bool

p24 :: t -> Bool

p25 :: (Eq a, Num a) => (a -> a) -> Bool

Exapmles for Baire -> Nat

fiNNN :: K (Baire -> Nat)

Existential quantifier for the set of zwei functions in Baire -> Nat.

fiImage :: K (Baire -> Nat)

Existential quantifier for continuous image of Cantor space in Baire -> Nat.

p31 :: (Baire -> Nat) -> Bool

Some predicates on Baire -> Nat.

p32 :: (Baire -> Nat) -> Bool

Some predicates on Baire -> Nat.

p33 :: (Baire -> Nat) -> Bool

Some predicates on Baire -> Nat.

p34 :: (Baire -> Nat) -> Bool

Some predicates on Baire -> Nat.