SelectionFunctions.NQueens

Description

We have a N x N chess board. Two players take turns placing queens on the board. On i-th turn a queen must be placed in i-th column so that previously placed queens are not attacking it. This module can be used to calculate optimal moves using selection functions implemented as described in paper What Sequential Games, the Tychonoff Theorem and the Double-Negation Shift have in Common, 2010, by Martin Escardo.

Documentation

n :: Int

Board size.

type Move = Int

Move is the number of column in which a queen is placed.

type Position = (Int, Int)

Coordinates on the board.

type R = Bool

Type of truth values that represent game outcomes.

attacks :: Position -> Position -> Bool

Call attacks p1 p2 checks if a queen in position p1 attacks position p2.

valid :: [Position] -> Bool

Function that checks if placement of queens on given positions is valid.

p :: [Move] -> R

Predicate on move sequences which tests the validity of resulting queen placement.

epsilons :: [[Move] -> J R Move]

List of history depentent selection functions of moves, where k-th selection function returns optimal move on k-th turn.

optimalPlay :: [Move]

Optimal play for NQueens.

optimalOutcome :: R

Optimal outcome of NQueens

optimalStrategy :: [Move] -> Move

Returns an optimal move, given previous moves.

delete :: Ord x => x -> [x] -> [x]

Function that removes an element from an ordered list.

setMinus :: Ord x => [x] -> [x] -> [x]

Call setMinus xs ys returns a list with elements of xs that are not contained in ys.