SelectionFunctions.Paper

Description

Module implements functions and examples described in paper What Sequential Games, the Tychonoff Theorem and the Double-Negation Shift have in Common, 2010, by Martin Escardo.

Documentation

type J r x = (x -> r) -> x

Selection functions on type x, where r is a type of generalized truth values.

type K r x = (x -> r) -> r

Generalized quantifiers over the type x, where r is a type of generalized truth values.

overline :: J r x -> K r x

Function mapping selection functions to quantifiers.

find :: [x] -> J Bool x

Finds an element of non-empty finite list that satisfies a predicate. If no such element exists it returns the last element.

findnot :: [x] -> J Bool x

Finds an element of non-empty finite list that doesn't satisfy a predicate. If no such element exists it returns the last element.

forsome :: [x] -> K Bool x

Existential quantifier over non-empty finite lists.

forevery :: [x] -> K Bool x

Universal quantifier over non-empty finite lists. The two definitions are equivalent.

argsupBool :: J Int Bool

Returns argsup of predicate Bool -> Int.

supBool :: K Int Bool

Returns sup of predicates Bool -> Int.

findBool :: J Bool Bool

Function find on a list [True, False].

argsup :: [x] -> J Int x

More efficient implementation of function argsup' if range of the predicate is -1,0,1. It avoids re-evaluations of p and stops when the maximum value is reached.

arginf :: [x] -> J Int x

Returns arginf of predicate x -> Int whose range is -1,0,1. It avoids re-evaluations of p and stops when the minimum value is reached.

otimes :: J r x0 -> J r x1 -> J r (x0, x1)

Binary product of selection functions.

bigotimes :: [J r x] -> J r [x]

Iterated product of a list of selection functions.

dotimes :: J r x -> (x -> J r [x]) -> J r [x]

History dependent binary product of selection functions.

dbigotimes :: [[x] -> J r x] -> J r [x]

History dependent iterated product of selection functions.

bigunion :: J r (J r x) -> J r x

Multiplication in monad J r.

mcons :: Monad m => m x -> m [x] -> m [x]

Generalization of function otimes' which works on any monad.

sequence' :: Monad m => [m x] -> m [x]

Alternative definition of the standard prelude function sequence, which is a generalization of function bigotimes to any monad.

hsequence :: Monad m => [[x] -> m x] -> m [x]

Generalization of function dbigotimes to any monad.

data Id x

Identity monad on type x.

fibonacci :: [Integer]

Computing Fibonacci numbers using hsequence on type Id Integer.