SelectionFunctions.TicTacToe

Description

The board and moves are 0 1 2 3 4 5 6 7 8 R is the set 3 = (-1,0,1) assumption: player X starts

This module can be used to calculate optimal moves using selection functions implemented as described in paper What Sequential Games, the Tychonoff Theorem and the Double-Negation Shift have in Common, 2010, by Martin Escardo.

Documentation

type R = Int

Type of truth values that represent game outcomes.

type Move = Int

Type used to describe player moves.

type Board = ([Move], [Move])

Type used to describe board state. Each list represents sequence of moves of a player.

data Player

Type of players - X and O.

Constructors

X

O

wins :: [Move] -> Bool

Function that checks if the game ends given after a given sequence of moves.

value :: Board -> R

Function that computes outcome of a board.

outcome :: Player -> [Move] -> Board -> Board

Function that takes the current player and a list of moves of both players and computes final state of the board.

p :: [Move] -> R

Predicate on lists of moves, that computes outcome of a game given a sequence of moves.

epsilons :: [[Move] -> J R Move]

List of history depentent selection functions of moves, where n-th selection function returns optimal move on n-th turn.

optimalPlay :: [Move]

Optimal moves for Tic Tac Toe.

optimalOutcome :: R

Optimal outcome of Tic Tac Toe.

optimalStrategy :: [Move] -> Move

Returns an optimal move, given previous moves.

contained :: Ord x => [x] -> [x] -> Bool

Call contained xs ys checks if an ordered list ys contains all elements of an ordered list xs.

someContained :: Ord x => [[x]] -> [x] -> Bool

Call someContained xss ys returns true iff some list xs in xss is contained in ys.

insert :: Ord x => x -> [x] -> [x]

Function that inserts an element in an ordered list so that it remains ordered.

delete :: Ord x => x -> [x] -> [x]

Function that removes an element from an ordered list.

setMinus :: Ord x => [x] -> [x] -> [x]

Call setMinus xs ys returns a list with elements of xs that are not contained in ys.